double

+++ MineSweeperSolver/BinomialHelper.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "BigInteger.h"
+#include "stdafx.h"
 
 extern "C" DLL_API void CacheBinomials(int n, int m);
-DLL_API BigInteger Binomial(int n, int m);
+DLL_API double Binomial(int n, int m);
